optimize `behav_eq!`
use Float32 in `behav_eq!`
give `behav_eq!` the number of trajectories.
Change `Float64` in fitness functions for ease.

+++ CoopConflGameStructs.jl
@@ -10,7 +10,7 @@ mutable struct Simulation_Parameters
     T0::Float64
     #popgen params
     gmax::Int64  # max generations
-    tmax::Float64  # max timespan for ODE
+    tmax::Float32  # max timespan for ODE
     N::Int64  # population size
     v::Float64  # synergy
     u::Float64  # mutation rate
